+++ buildSrc/src/main/kotlin/dependency/network/Coroutines.kt
@@ -6,5 +6,5 @@ object Coroutines : Dependency {
 
     override val group = "org.jetbrains.kotlinx"
     override val artifact = "kotlinx-coroutines"
-    override val version = "1.4.0"
+    override val version = "1.4.2"
 }