+++ instantsearch-android/src/main/java/com/algolia/instantsearch/helper/searcher/Extensions.kt
@@ -4,17 +4,17 @@ import com.algolia.instantsearch.core.InstantSearch
 import com.algolia.instantsearch.helper.searcher.internal.osVersion
 import com.algolia.search.dsl.requestOptions
 import com.algolia.search.model.Attribute
-import com.algolia.search.model.search.Query
+import com.algolia.search.model.params.CommonSearchParameters
 import com.algolia.search.transport.RequestOptions
 import io.ktor.http.HttpHeaders
 
-public fun Query.addFacet(vararg attribute: Attribute) {
+public fun CommonSearchParameters.addFacet(vararg attribute: Attribute) {
     facets = facets.orEmpty().toMutableSet().also {
         it += attribute
     }
 }
 
-public fun Query.removeFacet(attribute: Attribute) {
+public fun CommonSearchParameters.removeFacet(attribute: Attribute) {
     facets = facets.orEmpty().toMutableSet().also {
         it -= attribute
     }