+++ instantsearch-android/src/main/java/com/algolia/instantsearch/helper/stats/internal/StatsConnectionSearcher.kt
@@ -2,13 +2,13 @@ package com.algolia.instantsearch.helper.stats.internal
 
 import com.algolia.instantsearch.core.Callback
 import com.algolia.instantsearch.core.connection.ConnectionImpl
-import com.algolia.instantsearch.helper.searcher.SearcherSingleIndex
+import com.algolia.instantsearch.helper.searcher.SearcherIndex
 import com.algolia.instantsearch.helper.stats.StatsViewModel
 import com.algolia.search.model.response.ResponseSearch
 
 internal data class StatsConnectionSearcher(
     private val viewModel: StatsViewModel,
-    private val searcher: SearcherSingleIndex,
+    private val searcher: SearcherIndex<*>,
 ) : ConnectionImpl() {
 
     private val updateResponse: Callback<ResponseSearch?> = { response ->